# Add optimized method lookup to avoid creating intermediate method objects