--- a/vm/src/builtins/object.rs
+++ b/vm/src/builtins/object.rs
@@ -240,7 +240,7 @@ impl PyBaseObject {
 
     #[pymethod(name = "__getattribute__")]
     #[pyslot]
-    fn getattro(obj: PyObjectRef, name: PyStrRef, vm: &VirtualMachine) -> PyResult {
+    pub(crate) fn getattro(obj: PyObjectRef, name: PyStrRef, vm: &VirtualMachine) -> PyResult {
         vm_trace!("object.__getattribute__({:?}, {:?})", obj, name);
         vm.generic_getattribute(obj, name)
     }