--- a/vm/src/macros.rs
+++ b/vm/src/macros.rs
@@ -30,7 +30,7 @@ macro_rules! py_class {
             $($crate::py_class!(@extract_slots($ctx, &mut slots, $name, $value));)*
             let py_class = $ctx.new_class($class_name, $class_base, slots);
             $($crate::py_class!(@extract_attrs($ctx, &py_class, $name, $value));)*
-            $ctx.add_tp_new_wrapper(&py_class);
+            $ctx.add_slot_wrappers(&py_class);
             py_class
         }
     };
@@ -55,7 +55,7 @@ macro_rules! extend_class {
         $(
             $class.set_str_attr($name, $value);
         )*
-        $ctx.add_tp_new_wrapper(&$class);
+        $ctx.add_slot_wrappers(&$class);
     };
 }