… Lib/test/test_fileio.py
… Lib/test/test_with.py
… vm/src/builtins/pytype.rs
… vm/src/builtins/weakproxy.rs
… vm/src/cformat.rs
… vm/src/format.rs
… vm/src/frame.rs
… vm/src/function.rs
… vm/src/pyobject.rs
… vm/src/pyobjectrc.rs
… vm/src/slots.rs
… vm/src/stdlib/io.rs
… vm/src/stdlib/thread.rs
… vm/src/vm.rs