Sort of a follow-up to #2476

Benchmarks (without `threading`):

```
Benchmark #1: ./rustpython-norm benchmarks/pystone.py
  Time (mean ± σ):      2.699 s ±  0.124 s    [User: 2.678 s, System: 0.004 s]
  Range (min … max):    2.452 s …  2.870 s    10 runs
 
Benchmark #2: ./rustpython-opt benchmarks/pystone.py
  Time (mean ± σ):      2.366 s ±  0.120 s    [User: 2.349 s, System: 0.007 s]
  Range (min … max):    2.141 s …  2.581 s    10 runs
 
Summary
  './rustpython-opt benchmarks/pystone.py' ran
    1.14 ± 0.08 times faster than './rustpython-norm benchmarks/pystone.py'
```