@coolreader18 2021-03-14
I'm also gonna add the GET_METHOD/CALL_METHOD instructions, but idk if I should do it here, since it's already a pretty big changeset.